角色管理更新、删除

+++ erp/erp-service-impl/src/main/java/com/kaishengit/erp/service/impl/RolePermissionServiceImpl.java
@@ -3,12 +3,15 @@ package com.kaishengit.erp.service.impl;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Collections2;
 import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 import com.kaishengit.erp.entity.*;
 import com.kaishengit.erp.exception.ServiceException;
 import com.kaishengit.erp.mapper.PermissionMapper;
 import com.kaishengit.erp.mapper.RoleMapper;
 import com.kaishengit.erp.mapper.RolePermissionMapper;
 import com.kaishengit.erp.service.RolePermissionService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +20,7 @@ import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author jinjianghao
@@ -24,6 +28,8 @@ import java.util.List;
  */
 @Service
 public class RolePermissionServiceImpl implements RolePermissionService {
+
+    private Logger logger = LoggerFactory.getLogger(RolePermissionServiceImpl.class);
 
     @Autowired
     private PermissionMapper permissionMapper;
@@ -64,7 +70,7 @@ public class RolePermissionServiceImpl implements RolePermissionService {
      * @return
      */
     @Override
-    public List<Permission> findList() {
+    public List<Permission> findAllPermission() {
         PermissionExample permissionExample = new PermissionExample();
 
         List<Permission> permissionList = permissionMapper.selectByExample(permissionExample);
@@ -163,7 +169,11 @@ public class RolePermissionServiceImpl implements RolePermissionService {
             throw new ServiceException("该权限正在被使用，不能删除");
         }
 
-        permissionMapper.deleteByPrimaryKey(id);
+        Permission permission = permissionMapper.selectByPrimaryKey(id);
+        if(permission != null) {
+            permissionMapper.deleteByPrimaryKey(id);
+            logger.info("删除权限 {}", permission);
+        }
     }
 
     @Override
@@ -171,6 +181,98 @@ public class RolePermissionServiceImpl implements RolePermissionService {
         // 设置更新时间
         permission.setUpdateTime(new Date());
         permissionMapper.updateByPrimaryKeySelective(permission);
+        logger.info("更新权限 {}",permission);
+    }
+
+    /**
+     * 根据id删除角色
+     *
+     * @param id
+     * @throws ServiceException
+     */
+    @Override
+    @Transactional(rollbackFor = RuntimeException.class)
+    public void delRoleById(Integer id) throws ServiceException {
+        //查询是否被账号引用，如果引用则不能删除
+        //  TODO
+
+        //删除角色和权限的关系记录
+        RolePermissionExample rolePermissionExample = new RolePermissionExample();
+        rolePermissionExample.createCriteria().andRoleIdEqualTo(id);
+
+        rolePermissionMapper.deleteByExample(rolePermissionExample);
+        //删除角色
+        Role role = roleMapper.selectByPrimaryKey(id);
+        if(role != null) {
+            roleMapper.deleteByPrimaryKey(id);
+            logger.info("删除角色 {}",role);
+        }
+    }
+
+    /**
+     * 根据id查找角色附带角色的权限列表
+     * @return role
+     */
+    @Override
+    public Role findRoleWithPermission(Integer id) {
+        Role role = roleMapper.findByIdWithPermission(id);
+        return role;
+    }
+
+    /**
+     * 在编辑页面判断当前权限的复选框是否被checked
+     * @param rolePermissionList 当前角色拥有的权限
+     * @return 有顺序的map集合，如果被选择则value为true
+     */
+    @Override
+    public Map<Permission, Boolean> permissionBooleanMap(List<Permission> rolePermissionList) {
+        // 获得所有的权限列表
+        List<Permission> permissionList = findAllPermission();
+
+        // 获得有序的map集合
+        Map<Permission,Boolean> resultMap = Maps.newLinkedHashMap();
+
+        for(Permission permission : permissionList) {
+            boolean flag = false;
+            for(Permission rolePermission : rolePermissionList) {
+                if(permission.getId().equals(rolePermission.getId())) {
+                    flag = true;
+                    break;
+                }
+            }
+            resultMap.put(permission,flag);
+        }
+        return resultMap;
+    }
+
+    /**
+     * 更新角色
+     *
+     * @param role 角色对象
+     * @param permissionIds 角色拥有的新的权限id
+     */
+    @Override
+    @Transactional(rollbackFor = RuntimeException.class)
+    public void editRole(Role role, Integer[] permissionIds) {
+
+        //将角色原有和权限的对应关系删除
+        RolePermissionExample rolePermissionExample = new RolePermissionExample();
+        rolePermissionExample.createCriteria().andRoleIdEqualTo(role.getId());
+
+        rolePermissionMapper.deleteByExample(rolePermissionExample);
+        //重建角色和权限的对应关系
+        for(Integer permissionId : permissionIds) {
+            RolePermission rolePermission = new RolePermission();
+            rolePermission.setPermissionId(permissionId);
+            rolePermission.setRoleId(role.getId());
+
+            rolePermissionMapper.insert(rolePermission);
+        }
+        //修改角色对象
+        roleMapper.updateByPrimaryKeySelective(role);
+
+        logger.info("修改角色 {}",role);
+
     }
